// src/app/Modules/salary/Model/salary.model.ts
export class Salary {
    public employeeId: string ='';
    public employeeName: string ='';
    public departmentName: string ='';
    public designationId: string ='';
    public designationName: string ='';
    public status: string ='';
    public departmentId:string='';
  responseModel: Salary[];
  value: Salary;
  activeStatus: string;
  customerId: string;
  districtId: string;
  pageSize?: number;
  page?: number;
  count?:number
  }
  